+++ medic/src/cli/app.rs
@@ -80,6 +80,7 @@ impl From<ManifestArgs> for Flags {
     Self {
       auto_apply_remedy: args.apply_remedies,
       interactive: args.interactive,
+      ..Self::default()
     }
   }
 }